+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 - `[blocksync]` validate blocksync response sender and signature count
   ([\#5860](https://github.com/cometbft/cometbft/pull/5860))
-- `[blocksync]` test(blocksync): fix flaky `TestBlockPoolBasic` deadlock under `-race`
+- `[blocksync]` fix flaky `TestBlockPoolBasic` deadlock under `-race`
   ([\#5867](https://github.com/cometbft/cometbft/pull/5867))
 
 ### FEATURES